Create libraryclass
This is a code of a class created for a library system for lending, returning, donating and displaying list of book. there are four functions for each of them. class has two instance variables taken as argument 1. List of all book in the form of py dictionary and 2. the value in key-value pairs is a string which shows the available status of book. this string is given as argument in class